refactor: merge `AdvancedSearchPresenter` into its only subclass
